# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FormFlow API - AI Martial Arts Form Trainer
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile (name, camera side preference)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin only)
      parameters:
        - in: query
          name: role
          schema:
            type: string
        - in: query
          name: dojo_id
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{userId}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user role/dojo (admin only)
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}/sessions:
    get:
      operationId: getUserSessions
      tags: [users]
      summary: Get session history for a user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"

  # ─── Dojos ────────────────────────────────────────────────────────────────
...
  /stripe/checkout:
    post:
      operationId: createCheckoutSession
      tags: [stripe]
      summary: Create a Stripe Checkout session for student subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResult"
  /stripe/portal:
    post:
      operationId: createPortalSession
      tags: [stripe]
      summary: Create Stripe Customer Portal session for dojo billing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortalInput"
      responses:
        "200":
          description: Portal URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalResult"
components:
  schemas:
...
    User:
      type: object
      required: [id, email, role]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
        dojo_id:
          type: ["string", "null"]
        stripe_customer_id:
          type: ["string", "null"]
        subscription_status:
          type: string
        camera_side:
          type: ["string", "null"]
        created_at:
          type: string
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        camera_side:
          type: string
    UserUpdate:
      type: object
      properties:
        role:
          type: string
        dojo_id:
          type: ["string", "null"]
...
    Session:
      type: object
      required: [id, user_id, form_id]
      properties:
        id:
          type: string
        user_id:
          type: string
        form_id:
          type: string
        form_name:
          type: ["string", "null"]
        total_score:
          type: ["number", "null"]
        postures_completed:
          type: integer
        started_at:
          type: string
        completed_at:
          type: ["string", "null"]
...
    CheckoutInput:
      type: object
      required: [price_type]
      properties:
        price_type:
          type: string
        dojo_id:
          type: ["string", "null"]
        success_url:
          type: string
        cancel_url:
          type: string
    CheckoutResult:
      type: object
      required: [url]
      properties:
        url:
          type: string
    PortalInput:
      type: object
      required: [dojo_id]
      properties:
        dojo_id:
          type: string
        return_url:
          type: string
    PortalResult:
      type: object
      required: [url]
      properties:
        url:
          type: string